+++ src/edges_cal/cal_coefficients.py
@@ -28,221 +28,12 @@ from matplotlib import pyplot as plt
 from scipy.interpolate import InterpolatedUnivariateSpline as Spline
 
 from . import loss, s11
-from . import modelling as mdl
 from . import receiver_calibration_func as rcf
 from . import reflection_coefficient as rc
 from .cached_property import cached_property, safe_property
+from .loss import HotLoadCorrection
 from .spectra import LoadSpectrum
-from .tools import FrequencyRange, bin_array, get_data_path
-
-
-@hickleable()
-@attr.s(kw_only=True)
-class HotLoadCorrection:
-    """
-    Corrections for the hot load.
-
-    Measurements required to define the HotLoad temperature, from Monsalve et al.
-    (2017), Eq. 8+9.
-
-    Parameters
-    ----------
-    path
-        Path to a file containing measurements of the semi-rigid cable reflection
-        parameters. A preceding colon (:) indicates to prefix with DATA_PATH.
-        The default file was measured in 2015, but there is also a file included
-        that can be used from 2017: ":semi_rigid_s_parameters_2017.txt".
-    f_low, f_high
-        Lowest/highest frequency to retain from measurements.
-    n_terms
-        The number of terms used in fitting S-parameters of the cable.
-    """
-
-    freq: FrequencyRange = attr.ib()
-    raw_s11: np.ndarray = attr.ib(eq=attr.cmp_using(eq=np.array_equal))
-    raw_s12s21: np.ndarray = attr.ib(eq=attr.cmp_using(eq=np.array_equal))
-    raw_s22: np.ndarray = attr.ib(eq=attr.cmp_using(eq=np.array_equal))
-
-    model: mdl.Model = attr.ib(mdl.Polynomial(n_terms=21))
-    complex_model: type[mdl.ComplexRealImagModel] | type[
-        mdl.ComplexMagPhaseModel
-    ] = attr.ib(mdl.ComplexMagPhaseModel)
-    use_spline: bool = attr.ib(False)
-    model_method: str = attr.ib("lstsq")
-
-    @classmethod
-    def from_file(
-        cls,
-        path: tp.PathLike = ":semi_rigid_s_parameters_WITH_HEADER.txt",
-        f_low: tp.FreqType = 0 * un.MHz,
-        f_high: tp.FreqType = np.inf * un.MHz,
-        set_transform_range: bool = True,
-        **kwargs,
-    ):
-        """Instantiate the HotLoadCorrection from file.
-
-        Parameters
-        ----------
-        path
-            Path to the S-parameters file.
-        f_low, f_high
-            The min/max frequencies to use in the modelling.
-        """
-        path = get_data_path(path)
-
-        data = np.genfromtxt(path)
-        freq = FrequencyRange(data[:, 0] * un.MHz, f_low=f_low, f_high=f_high)
-
-        data = data[freq.mask]
-
-        if data.shape[1] == 7:  # Original file from 2015
-            data = data[:, 1::2] + 1j * data[:, 2::2]
-        elif data.shape[1] == 6:  # File from 2017
-            data = np.array(
-                [
-                    data[:, 1] + 1j * data[:, 2],
-                    data[:, 3],
-                    data[:, 4] + 1j * data[:, 5],
-                ]
-            ).T
-
-        model = kwargs.pop(
-            "model",
-            mdl.Polynomial(
-                n_terms=21,
-                transform=mdl.UnitTransform(
-                    range=(freq.min.to_value("MHz"), freq.max.to_value("MHz"))
-                ),
-            ),
-        )
-
-        if hasattr(model.transform, "range") and set_transform_range:
-            model = attr.evolve(
-                model,
-                transform=attr.evolve(
-                    model.transform,
-                    range=(freq.min.to_value("MHz"), freq.max.to_value("MHz")),
-                ),
-            )
-
-        return cls(
-            freq=freq,
-            raw_s11=data[:, 0],
-            raw_s12s21=data[:, 1],
-            raw_s22=data[:, 2],
-            model=model,
-            **kwargs,
-        )
-
-    def _get_model(self, raw_data: np.ndarray, **kwargs):
-        model = self.complex_model(self.model, self.model)
-        return model.fit(xdata=self.freq.freq, ydata=raw_data, method=self.model_method)
-
-    def _get_splines(self, data):
-        if self.complex_model == mdl.ComplexRealImagModel:
-            return (
-                Spline(self.freq.freq.to_value("MHz"), np.real(data)),
-                Spline(self.freq.freq.to_value("MHz"), np.imag(data)),
-            )
-        return (
-            Spline(self.freq.freq.to_value("MHz"), np.abs(data)),
-            Spline(self.freq.freq.to_value("MHz"), np.angle(data)),
-        )
-
-    def _ev_splines(self, splines):
-        rl, im = splines
-        if self.complex_model == mdl.ComplexRealImagModel:
-            return lambda freq: rl(freq) + 1j * im(freq)
-        return lambda freq: rl(freq) * np.exp(1j * im(freq))
-
-    @cached_property
-    def s11_model(self):
-        """The reflection coefficient."""
-        if not self.use_spline:
-            return self._get_model(self.raw_s11)
-        splines = self._get_splines(self.raw_s11)
-        return self._ev_splines(splines)
-
-    @cached_property
-    def s12s21_model(self):
-        """The transmission coefficient."""
-        if not self.use_spline:
-            return self._get_model(self.raw_s12s21)
-        splines = self._get_splines(self.raw_s12s21)
-        return self._ev_splines(splines)
-
-    @cached_property
-    def s22_model(self):
-        """The reflection coefficient from the other side."""
-        if not self.use_spline:
-            return self._get_model(self.raw_s22)
-        splines = self._get_splines(self.raw_s22)
-        return self._ev_splines(splines)
-
-    def power_gain(self, freq: tp.FreqType, hot_load_s11: s11.LoadS11) -> np.ndarray:
-        """
-        Calculate the power gain.
-
-        Parameters
-        ----------
-        freq : np.ndarray
-            The frequencies.
-        hot_load_s11 : :class:`LoadS11`
-            The S11 of the hot load.
-
-        Returns
-        -------
-        gain : np.ndarray
-            The power gain as a function of frequency.
-        """
-        assert isinstance(hot_load_s11, s11.LoadS11), "hot_load_s11 must be a LoadS11"
-        assert (
-            hot_load_s11.load_name == "hot_load"
-        ), "hot_load_s11 must be a hot_load s11"
-
-        return self.get_power_gain(
-            {
-                "s11": self.s11_model(freq.to_value("MHz")),
-                "s12s21": self.s12s21_model(freq.to_value("MHz")),
-                "s22": self.s22_model(freq.to_value("MHz")),
-            },
-            hot_load_s11.s11_model(freq.to_value("MHz")),
-        )
-
-    @staticmethod
-    def get_power_gain(
-        semi_rigid_sparams: dict, hot_load_s11: np.ndarray
-    ) -> np.ndarray:
-        """Define Eq. 9 from M17.
-
-        Parameters
-        ----------
-        semi_rigid_sparams : dict
-            A dictionary of reflection coefficient measurements as a function of
-            frequency for the semi-rigid cable.
-        hot_load_s11 : array-like
-            The S11 measurement of the hot_load.
-
-        Returns
-        -------
-        gain : np.ndarray
-            The power gain.
-        """
-        rht = rc.gamma_de_embed(
-            semi_rigid_sparams["s11"],
-            semi_rigid_sparams["s12s21"],
-            semi_rigid_sparams["s22"],
-            hot_load_s11,
-        )
-
-        return (
-            np.abs(semi_rigid_sparams["s12s21"])
-            * (1 - np.abs(rht) ** 2)
-            / (
-                (np.abs(1 - semi_rigid_sparams["s11"] * rht)) ** 2
-                * (1 - np.abs(hot_load_s11) ** 2)
-            )
-        )
+from .tools import FrequencyRange, bin_array
 
 
 @hickleable()
@@ -273,7 +64,7 @@ class Load:
     def loss_model(self):
         """The loss model as a callable function of frequency."""
         if isinstance(self._loss_model, HotLoadCorrection):
-            return partial(self._loss_model.power_gain, hot_load_s11=self.reflections)
+            return self._loss_model.power_gain
         return self._loss_model
 
     @property
@@ -918,7 +709,7 @@ class CalibrationObservation:
         reflection_kwargs = reflection_kwargs or {}
         spec_kwargs = spec_kwargs or {}
 
-        spec_kwargs["freq_bin_size"] = self.freq.bin_size
+        spec_kwargs["freq_bin_size"] = self.metadata.get("freq_bin_size", 1)
         spec_kwargs["t_load"] = self.open.spectrum.t_load
         spec_kwargs["t_load_ns"] = self.open.spectrum.t_load_ns
 
@@ -927,14 +718,15 @@ class CalibrationObservation:
                 "frequency_smoothing"
             ]
 
-        spec_kwargs["f_low"] = self.freq._f_low
-        spec_kwargs["f_high"] = self.freq._f_high
+        spec_kwargs["f_low"], spec_kwargs["f_high"] = self.open.spectrum.metadata[
+            "pre_smooth_freq_range"
+        ]
 
         return Load.from_io(
             io_obj=io_obj,
             load_name=load_name,
-            f_low=self.freq.post_bin_f_low,
-            f_high=self.freq.post_bin_f_high,
+            f_low=self.freq._f_low,
+            f_high=self.freq._f_high,
             reflection_kwargs=reflection_kwargs,
             spec_kwargs=spec_kwargs,
         )
